refactor(about): switch to slots over passing props
